separated transport_worker into a separate file
