+++ axon/worker.py
@@ -1,123 +1,18 @@
 # the bit that listens idly and serves RPCs
 # and executor is the thing that takes a function, and executes it in a thread/process and returns the result as response or even another HTTP req entirely
 
-from .utils import deserialize, serialize, get_self_ip, overwrite
+from .utils import overwrite
 from .config import comms_config, default_rpc_endpoint, default_service_config, default_service_depth
+from .transport_worker import HTTPTransportWorker
 
-from flask import Flask
-from flask import request as route_req
 from copy import copy
-from threading import Lock
-from concurrent.futures import ProcessPoolExecutor as PPE
-from threading import Thread
-import inspect
-import random
-import string
-import traceback
-import sys
-import logging
 import pickle, cloudpickle
 import asyncio
 
-# the ip address of the worker
-ip_addr = get_self_ip()
-cli = sys.modules['flask.cli']
-cli.show_server_banner = lambda *x: None
-log = logging.getLogger('werkzeug')
-log.disabled = True
-# the app that listens for incomming http requests
-app = Flask(__name__)
-name = 'worker'
-
-loop, event_loop_thread = None, None
-def start_event_loop_thread():
-	global loop, event_loop_thread
-	loop = asyncio.new_event_loop()
-	asyncio.set_event_loop(loop)
-	event_loop_thread = Thread(target=loop.run_forever, daemon=True)
-	event_loop_thread.start()
-
-inline_lock = Lock()
-def invoke_RPC(target_fn, param_str, in_parallel=True):
-	global loop, event_loop_thread
-
-	if isinstance(target_fn, bytes) or isinstance(target_fn, str):
-		target_fn = pickle.loads(target_fn)
-
-	args, kwargs = deserialize(param_str)
-
-	return_object = {
-		'errcode': 0,
-		'result': None,
-	}
-
-	try:
-		result = None
-
-		if not in_parallel:
-			with inline_lock:
-				result = target_fn(*args, **kwargs)
-		else:
-			result = target_fn(*args, **kwargs)
-
-		if inspect.iscoroutine(result):
-			if (event_loop_thread == None) or not (event_loop_thread.is_alive()):
-				start_event_loop_thread()
-				
-			result = asyncio.run_coroutine_threadsafe(result, loop).result()
-
-		return_object['result'] = result
-
-	except:
-		return_object['errcode'] = 1
-		return_object['result'] = (traceback.format_exc(), sys.exc_info()[1])
-
-	return serialize(return_object)
-
-def register_RPC(fn, **configuration):
-
-	if not 'name' in configuration:
-		configuration['name'] = fn.__name__
-
-	executor = configuration['executor']
-
-	if isinstance(executor, PPE):
-		fn = cloudpickle.dumps(fn)
-
-	def route_fn():
-		param_str = route_req.form['msg']
-		future = executor.submit(invoke_RPC, fn, param_str)
-		return future.result()
-
-	# flask requires that each route function has a unique name
-	route_fn.__name__ = ''.join(random.choices(string.ascii_letters, k=10))
-	endpoint = '/'+configuration['endpoint_prefix']+configuration['name']
-	app.route(endpoint, methods=['POST'])(route_fn)
-
-class HTTPTransportWorker():
-
-	def __init__(self):
-		pass
-
-	def register_RPC(self, fn, **configuration):
-
-		if not 'name' in configuration:
-			configuration['name'] = fn.__name__
-
-		executor = configuration['executor']
-
-		if isinstance(executor, PPE):
-			fn = cloudpickle.dumps(fn)
-
-		def route_fn():
-			param_str = route_req.form['msg']
-			future = executor.submit(invoke_RPC, fn, param_str)
-			return future.result()
-
-		# flask requires that each route function has a unique name
-		route_fn.__name__ = ''.join(random.choices(string.ascii_letters, k=10))
-		endpoint = '/'+configuration['endpoint_prefix']+configuration['name']
-		app.route(endpoint, methods=['POST'])(route_fn)
+# before we factor out the HTTPTransportWorker, we need to use None as the default setting in default_service_config
+# this is because this file imports from config, and so config cannot import from this file
+if (default_service_config['tl'] == None):
+	default_service_config['tl'] = HTTPTransportWorker(comms_config.worker_port)
 
 def _get_profile():
 
@@ -149,10 +44,6 @@ class ServiceNode():
 		self.depth = depth
 
 		self.configuration = overwrite(default_service_config, configuration)
-		# before we factor out the HTTPTransportWorker, we need to use None as the default setting in default_service_config
-		# this is because this file imports from config, and so config cannot import from this file
-		if (self.configuration['tl'] == None):
-			self.configuration['tl'] = HTTPTransportWorker()
 
 		# iterates over members and either registers them as RPCs or recursively turns them into ServiceNodes
 		# for key, member in self.subject.__dict__.items():
@@ -211,7 +102,6 @@ class ServiceNode():
 		}
 
 		tl.register_RPC(fn, **child_config)
-		# register_RPC(fn, **child_config)
 
 		# remember the configuration
 		self.children[key] = child_config
@@ -250,12 +140,10 @@ def rpc(**configuration):
 
 	return add_to_RPC_node
 
-# starts the web app
-def init(wrkr_name='worker', port=comms_config.worker_port):
+def init(port=comms_config.worker_port):
 	global name
 
-	register_RPC(_get_profile, name='/_get_profile', executor=default_service_config['executor'], endpoint_prefix=default_service_config['endpoint_prefix'])
+	tl = default_service_config['tl']
+	tl.register_RPC(_get_profile, name='/_get_profile', executor=default_service_config['executor'], endpoint_prefix=default_service_config['endpoint_prefix'])
 
-	name = wrkr_name
-	# the web application that will serve the services and rpcs as routes
-	app.run(host='0.0.0.0', port=port)
+	tl.run()
